+++ src/confcom/azext_confcom/security_policy.py
@@ -12,7 +12,7 @@ from typing import Any, Dict, List, Tuple, Union
 from dataclasses import asdict
 
 from azext_confcom.lib.arm_to_aci_policy_spec import arm_to_aci_policy_spec
-from azext_confcom import (config, os_util)
+from azext_confcom import config, os_util
 from azext_confcom.container import ContainerImage, UserContainerImage
 from azext_confcom.errors import eprint
 from azext_confcom.fragment_util import sanitize_fragment_fields
@@ -259,7 +259,7 @@ class AciPolicy:  # pylint: disable=too-many-instance-attributes
         return policy.validate(policy_content, sidecar_validation=True)
 
     # pylint: disable=too-many-locals
-    def validate(self, policy, sidecar_validation=False) -> Tuple[bool, Dict]:
+    def validate(self, container_policy_list, sidecar_validation=False) -> Tuple[bool, Dict]:
         """Utility method: general method to compare two policies.
         One being the current object and the other is passed in as a parameter.
 
@@ -270,8 +270,8 @@ class AciPolicy:  # pylint: disable=too-many-instance-attributes
         The minimum difference is used to match up the containers in the policy vs
         the containers in the ARM template. Afterwards, the differences are compiled
         and returned as a dictionary organized by container name."""
-        if not policy:
-            eprint("Policy is not in the expected form to validate against")
+        if not container_policy_list:
+            container_policy_list = []
 
         policy_str = self.get_serialized_output(
             OutputType.PRETTY_PRINT, rego_boilerplate=False
@@ -282,11 +282,11 @@ class AciPolicy:  # pylint: disable=too-many-instance-attributes
 
         policy_ids = [
             case_insensitive_dict_get(i, config.POLICY_FIELD_CONTAINERS_ID)
-            for i in policy
+            for i in container_policy_list
         ]
         policy_names = [
             case_insensitive_dict_get(i, config.POLICY_FIELD_CONTAINERS_NAME)
-            for i in policy
+            for i in container_policy_list
         ]
 
         for container in arm_containers:
@@ -320,7 +320,7 @@ class AciPolicy:  # pylint: disable=too-many-instance-attributes
             temp_diff_list = []
             for idx in set_idx:
                 temp_diff = {}
-                matching_policy_container = policy[idx]
+                matching_policy_container = container_policy_list[idx]
 
                 diff_values = get_container_diff(matching_policy_container, container)
                 # label the diff with the ID so it can be merged